fix(db_stats.py): Reorganize Prometheus query methods
	Prometheus queries related methods were not organized
	and were located in a specif test and not a common related file.
